Move ultrasonic classes/code into neopixel lib and app
